Changed project structure, so it follows Standard Directory Layout. Added default phrases.yml, so it's more intuitive to translate Fe. You don't need to put lower-case variables anymore in phrases.yml.
