+++ .gitignore
@@ -34,3 +34,4 @@ dependency-reduced-pom.xml
 *.ipr
 *.iws
 .idea/
+/.idea/
