+++ src/main/java/io/loyloy/fe/command/CommandType.java
