+++ src/main/java/io/loyloy/fe/command/commands/SetCommand.java
@@ -2,25 +2,23 @@ package io.loyloy.fe.command.commands;
 
 import io.loyloy.fe.Fe;
 import io.loyloy.fe.Phrase;
+import io.loyloy.fe.command.CommandType;
 import io.loyloy.fe.command.SubCommand;
 import io.loyloy.fe.database.Account;
 import org.bukkit.command.Command;
 import org.bukkit.command.CommandSender;
-import org.bukkit.entity.Player;
-import io.loyloy.fe.command.CommandType;
 
-public class DeductCommand extends SubCommand
+public class SetCommand extends SubCommand
 {
     private final Fe plugin;
 
-    public DeductCommand( Fe plugin )
+    public SetCommand( Fe plugin )
     {
-        super( "deduct", "fe.deduct", "deduct [name] [amount]", Phrase.COMMAND_DEDUCT, CommandType.CONSOLE );
+        super( "set", "fe.set", "set [name] [amount]", Phrase.COMMAND_SET, CommandType.CONSOLE );
 
         this.plugin = plugin;
     }
 
-    @SuppressWarnings( "deprecation" )
     public boolean onCommand( CommandSender sender, Command cmd, String commandLabel, String[] args )
     {
         if( args.length < 2 )
@@ -47,19 +45,18 @@ public class DeductCommand extends SubCommand
             return true;
         }
 
+        if( !victim.canReceive( money ) )
+        {
+            Phrase.MAX_BALANCE_REACHED.sendWithPrefix( sender, victim.getName() );
+            return true;
+        }
+
         String formattedMoney = plugin.getAPI().format( money );
 
-        victim.withdraw( money );
+        victim.setMoney( money );
 
-        Phrase.PLAYER_DEDUCT_MONEY.sendWithPrefix( sender, formattedMoney, victim.getName() );
-
-        Player receiverPlayer = plugin.getServer().getPlayerExact( victim.getName() );
-
-        if( receiverPlayer != null )
-        {
-            Phrase.PLAYER_DEDUCTED_MONEY.sendWithPrefix( receiverPlayer, formattedMoney, sender.getName() );
-        }
+        Phrase.PLAYER_SET_MONEY.sendWithPrefix( sender, victim.getName(), formattedMoney );
 
         return true;
     }
-}	
+}
