+++ src/main/resources/plugin.yml
@@ -1,7 +1,7 @@
 name: Fe
 main: io.loyloy.fe.Fe
-version: 0.9.1
-load: startup
+version: 0.9.2
+load: STARTUP
 
 commands:
   fe:
